// api/src/generated/prisma/models/Event.ts
export type EventModel = runtime.Types.Result.DefaultSelection<Prisma.$EventPayload>
// ...
export type $EventPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Event"
  objects: {
    user: Prisma.$UserPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    user_id: string
    google_event_id: string
    title: string
    summary: string | null
    start_time: Date
    end_time: Date
    created_at: Date
    is_all_day: boolean
  }, ExtArgs["result"]["event"]>
  composites: {}
}